requirements.txt
click==8.2.1
filelock==3.18.0
fsspec==2025.7.0
gensim==4.3.3
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
pandas==2.2.2
pillow==11.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.31
scipy==1.13.1
six==1.17.0
smart_open==7.3.0.post1
sympy==1.14.0
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
wrapt==1.17.2
